Add f2fs tuning for unencrypted devices
Why tune f2fs here in Magisk?

f2fs being a LFS file-system, needs to perform GC periodically.
Android handles this via vold's IdleMaint service.

The kernel throttles this behavior as it can affect foreground performance
via gc_urgent_sleep_time, which is set to 500ms by default.

Android overrides this from init.rc to 50ms, as vold is pretty clever when
it comes to idle detection and 500ms takes too long to GC all invalid
segments in time, which results in long-term performance degradation.

Unfortunately, due to how Android init works, the logic to write 50 to
gc_urgent_sleep_time breaks when the device is unencrypted.
(The fix [1] for this is expected to ship with Android 13.)

Moreover, f2fs-specific optimizations such as cp_interval and discard
limit also breaks.

When Android added support for f2fs platform-wise, encryption was enforced
to OEMs, so while this isn't a problem for production devices running
unmodified OS, it is for rooted users.

Many people end up decrypting their device (some are forced to do so as
unlocked bootloader forces users to use unencrypted userdata), and they
will be met with storage performance going noticeably slower as the time
goes on (typically within 1-2 months).

As Magisk is the go-to solution for rooting your device, it makes the best
sense to implement a workaround for this in Magisk itself without relying
on a separate module.

The values were taken directly from AOSP and weren't changed in years.

[1] https://android-review.googlesource.com/c/platform/system/core/+/1717435
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>