recreate every mountpoint under mirrors
This #6445 looks good because it can address many problems related to overlayfs.
For some reason, it causes bootloop for some device with rootdir recursively bind mounted to $(magisk --path)/.magisk/mirror. However it will be fixed if we setup mirror mount in usual way and then recursively bind mount its subtree. The code might look not good, maybe you can find a better way than this 😀
This PR and partial code is based on #6445
Credit: @yujincheng08

Co-Authored-By: LoveSy <shana@zju.edu.cn>